+++ export/multinode.go
@@ -3,12 +3,13 @@ package export
 import (
 	"context"
 	"fmt"
-	"path"
-
 	doguv2 "github.com/cloudogu/k8s-dogu-operator/v3/api/v2"
 	apiCorev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/watch"
 	corev1 "k8s.io/client-go/kubernetes/typed/core/v1"
+	"log/slog"
+	"time"
 )
 
 const (
@@ -30,19 +31,25 @@ type serviceClient interface {
 	Update(ctx context.Context, service *apiCorev1.Service, opts metav1.UpdateOptions) (*apiCorev1.Service, error)
 }
 
-type MultinodeExportModeProvider struct {
-	namespace     string
-	configMaps    configMaps
-	doguclient    doguClient
-	serviceclient serviceClient
+type endpointsClient interface {
+	Watch(ctx context.Context, opts metav1.ListOptions) (watch.Interface, error)
 }
 
-func NewMultinodeExportModeProvider(namespace string, configMaps configMaps, ecosystemDoguClient doguClient, serviceClient serviceClient) *MultinodeExportModeProvider {
+type MultinodeExportModeProvider struct {
+	namespace       string
+	configMaps      configMaps
+	doguclient      doguClient
+	serviceclient   serviceClient
+	endpointsclient endpointsClient
+}
+
+func NewMultinodeExportModeProvider(namespace string, configMaps configMaps, ecosystemDoguClient doguClient, serviceClient serviceClient, endpointsClient endpointsClient) *MultinodeExportModeProvider {
 	return &MultinodeExportModeProvider{
-		namespace:     namespace,
-		configMaps:    configMaps,
-		doguclient:    ecosystemDoguClient,
-		serviceclient: serviceClient,
+		namespace:       namespace,
+		configMaps:      configMaps,
+		doguclient:      ecosystemDoguClient,
+		serviceclient:   serviceClient,
+		endpointsclient: endpointsClient,
 	}
 }
 
@@ -54,7 +61,7 @@ func (m MultinodeExportModeProvider) GetExportDogu(ctx context.Context) (*doguEx
 
 	dE := &doguExport{
 		Dogu:         doguName,
-		VolumePath:   path.Join(dataVolumePath, doguName),
+		VolumePath:   dataVolumePath,
 		ExporterPort: int(port.Port),
 	}
 	return dE, nil
@@ -76,12 +83,48 @@ func (m MultinodeExportModeProvider) SetExportDogu(ctx context.Context, doguName
 		return nil, fmt.Errorf("failed to update exporter service: %w", err)
 	}
 
+	if err := m.waitForServiceEndpoints(ctx, service); err != nil {
+		return nil, fmt.Errorf("failed to wait for endpoints to update: %w", err)
+	}
+
 	dE := &doguExport{
 		Dogu:         doguName,
-		VolumePath:   path.Join(dataVolumePath, doguName),
+		VolumePath:   dataVolumePath,
 		ExporterPort: int(port.Port),
 	}
 	return dE, nil
+}
+
+func (m MultinodeExportModeProvider) waitForServiceEndpoints(ctx context.Context, service *apiCorev1.Service) error {
+	slog.Debug("start waiting for endpoints to update", "service", service.Name)
+	endpointsWatcher, err := m.endpointsclient.Watch(ctx, metav1.ListOptions{
+		FieldSelector: fmt.Sprintf("metadata.name=%s", service.Name),
+	})
+	if err != nil {
+		return fmt.Errorf("failed to start watch on endpoints: %w", err)
+	}
+	defer endpointsWatcher.Stop()
+
+	timeoutCh := time.After(time.Second * 10)
+	for {
+		select {
+		case event := <-endpointsWatcher.ResultChan():
+			if event.Type == watch.Error {
+				return fmt.Errorf("watch error occurred watching endpoints: %s", event.Object)
+			}
+			if ep, ok := event.Object.(*apiCorev1.Endpoints); ok {
+				// Check if any address is available
+				for _, subset := range ep.Subsets {
+					if len(subset.Addresses) > 0 {
+						slog.Debug("endpoints available", "service", service.Name, "subset", subset.Addresses[0].TargetRef.Name)
+						return nil
+					}
+				}
+			}
+		case <-timeoutCh:
+			return fmt.Errorf("timeout waiting for endpoints to update for service %s", service.Name)
+		}
+	}
 }
 
 func (m MultinodeExportModeProvider) GetExportMode(ctx context.Context) (*exportModeStatus, error) {